+++ module/utils.js
@@ -23,3 +23,16 @@ export function useAlternateButtonBehavior() {
 
   return useAlt;
 }
+
+//Function to remove any HTML markup from eg. item descriptions
+export function removeHtmlTags(str) {
+  // Replace any HTML tag ('<...>') by an empty string
+  // and then un-escape any HTML escape codes (eg. &lt;)
+  return htmlDecode(str.replace(/<.+?>/gi, ""));
+}
+
+// Stolen from https://stackoverflow.com/a/34064434/20043
+export function htmlDecode(input) {
+  var doc = new DOMParser().parseFromString(input, "text/html");
+  return doc.documentElement.textContent;
+}
